Schema: add option to preserve excess properties, closes #2008
