+++ .changeset/odd-bulldogs-knock.md
@@ -1,0 +1,5 @@
+---
+"@effect/schema": patch
+---
+
+add option to preserve excess properties, closes #2008
